+++ zcash_voting/src/wire_codec.rs
@@ -1300,7 +1300,8 @@ impl From<&VotingError> for VotingErrorView {
             endpoint: None,
         };
         match error {
-            VotingError::KeystoneSignatureConflict { bundle_index } => {
+            VotingError::KeystoneSignatureConflict { bundle_index }
+            | VotingError::DelegationReconciliationRequired { bundle_index, .. } => {
                 view.bundle_index = Some(*bundle_index);
             }
             VotingError::InsufficientEligibility {
@@ -1348,6 +1349,7 @@ impl From<crate::types::DelegationSetupField> for DelegationSetupFieldView {
             F::PaddedNoteSecrets => Self::PaddedNoteSecrets,
             F::PcztSighash => Self::PcztSighash,
             F::Tx1Effects => Self::Tx1Effects,
+            F::DelegationPczt => Self::DelegationPczt,
         }
     }
 }
@@ -1370,6 +1372,9 @@ impl From<VotingErrorKind> for VotingErrorKindView {
             VotingErrorKind::InsufficientEligibility => Self::InsufficientEligibility,
             VotingErrorKind::NoSpendableNotes => Self::NoSpendableNotes,
             VotingErrorKind::SetupAlreadyPersisted => Self::SetupAlreadyPersisted,
+            VotingErrorKind::DelegationReconciliationRequired => {
+                Self::DelegationReconciliationRequired
+            }
             VotingErrorKind::DbBusy => Self::DbBusy,
             VotingErrorKind::PirUnavailable => Self::PirUnavailable,
         }